Conflicts:
	fdb-sql-layer-core/src/main/java/com/foundationdb/qp/row/AbstractRow.java
	fdb-sql-layer-core/src/main/java/com/foundationdb/qp/storeadapter/NewRowBackedIndexRow.java

+++ fdb-sql-layer-core/src/test/java/com/foundationdb/server/test/it/qp/GroupScanIT.java
@@ -150,8 +150,8 @@ public class GroupScanIT extends OperatorITBase
         ValueSource v0 = row.value(0);
         ValueSource v1 = row.value(1);
         ValueSource v2 = row.value(2);
-        assertEquals(11L, v0.getInt64());
-        assertEquals(1L, v1.getInt64());
+        assertEquals(11L, v0.getInt32());
+        assertEquals(1L, v1.getInt32());
         assertEquals("ori", v2.getString());
     }
